Improved paid members change chart
ref https://linear.app/ghost/issue/FEA-473/tier-breakdown-in-analytics

- Moved "Change" out from the Paid members KPI chart so it's more discoverable
- Added Pie chart of cadence breakdown
